requirements.txt
fastapi
uvicorn
pydantic
redis
requests
pyyaml
sqlalchemy
pydantic-settings
